+++ tools/cosmovisor/args_test.go
@@ -609,14 +609,12 @@ func (s *argsTestSuite) TestGetConfigFromEnv() {
 			cfg, err := GetConfigFromEnv()
 			if tc.expectedErrCount == 0 {
 				assert.NoError(t, err)
-			} else {
-				if assert.Error(t, err) {
-					errCount := 1
-					if multi, isMulti := err.(*errors.MultiError); isMulti {
-						errCount = multi.Len()
-					}
-					assert.Equal(t, tc.expectedErrCount, errCount, "error count")
+			} else if assert.Error(t, err) {
+				errCount := 1
+				if multi, isMulti := err.(*errors.MultiError); isMulti {
+					errCount = multi.Len()
 				}
+				assert.Equal(t, tc.expectedErrCount, errCount, "error count")
 			}
 			assert.Equal(t, tc.expectedCfg, cfg, "config")
 		})